requirements.txt
python-dotenv==0.13.0
celery==4.4.2
requests==2.23.0
pymongo==3.10.1
redis==3.5.0
pytest==5.4.1
mediacloud-cliff==2.6.1
mediacloud==3.11.2
genderize==0.3.1
dogpile.cache==0.9.2
deco==0.5.2
